+++ src/main/java/com/easypost/service/ApiKeyService.java
@@ -5,7 +5,7 @@ import com.easypost.http.Requestor;
 import com.easypost.http.Requestor.RequestMethod;
 import com.easypost.model.ApiKey;
 import com.easypost.model.ApiKeys;
-import com.easypost.utils.Utilities;
+import com.easypost.utils.InternalUtilities;
 
 public class ApiKeyService {
     private final EasyPostClient client;
@@ -26,7 +26,7 @@ public class ApiKeyService {
      * @throws EasyPostException when the request fails.
      */
     public ApiKeys all() throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.classURL(ApiKey.class),
+        return Requestor.request(RequestMethod.GET, InternalUtilities.classURL(ApiKey.class),
             null, ApiKeys.class, client);
     }
 }