+++ src/main/java/com/easypost/service/CustomsItemService.java
@@ -4,7 +4,7 @@ import com.easypost.exception.EasyPostException;
 import com.easypost.http.Requestor;
 import com.easypost.http.Requestor.RequestMethod;
 import com.easypost.model.CustomsItem;
-import com.easypost.utils.Utilities;
+import com.easypost.utils.InternalUtilities;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -32,7 +32,7 @@ public class CustomsItemService {
         Map<String, Object> wrappedParams = new HashMap<String, Object>();
         wrappedParams.put("customs_item", params);
 
-        return Requestor.request(RequestMethod.POST, Utilities.classURL(CustomsItem.class),
+        return Requestor.request(RequestMethod.POST, InternalUtilities.classURL(CustomsItem.class),
                 wrappedParams, CustomsItem.class, client);
     }
 
@@ -44,7 +44,7 @@ public class CustomsItemService {
      * @throws EasyPostException when the request fails.
      */
     public CustomsItem retrieve(final String id) throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.instanceURL(CustomsItem.class, id),
+        return Requestor.request(RequestMethod.GET, InternalUtilities.instanceURL(CustomsItem.class, id),
                 null, CustomsItem.class, client);
     }
 }