+++ src/main/java/com/easypost/service/RefundService.java
@@ -5,7 +5,7 @@ import com.easypost.http.Requestor;
 import com.easypost.http.Requestor.RequestMethod;
 import com.easypost.model.Refund;
 import com.easypost.model.RefundCollection;
-import com.easypost.utils.Utilities;
+import com.easypost.utils.InternalUtilities;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -35,7 +35,7 @@ public class RefundService {
         Map<String, Object> wrappedParams = new HashMap<String, Object>();
         wrappedParams.put("refund", params);
 
-        Refund[] response = Requestor.request(RequestMethod.POST, Utilities.classURL(Refund.class),
+        Refund[] response = Requestor.request(RequestMethod.POST, InternalUtilities.classURL(Refund.class),
                 wrappedParams, Refund[].class, client);
 
         return Arrays.asList(response);
@@ -49,7 +49,7 @@ public class RefundService {
      * @throws EasyPostException when the request fails.
      */
     public Refund retrieve(final String id) throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.instanceURL(Refund.class, id), null, Refund.class,
+        return Requestor.request(RequestMethod.GET, InternalUtilities.instanceURL(Refund.class, id), null, Refund.class,
                 client);
     }
 
@@ -61,7 +61,7 @@ public class RefundService {
      * @throws EasyPostException when the request fails.
      */
     public RefundCollection all(final Map<String, Object> params) throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.classURL(Refund.class), params, RefundCollection.class,
-                client);
+        return Requestor.request(RequestMethod.GET, InternalUtilities.classURL(Refund.class), params,
+                RefundCollection.class, client);
     }
 }