+++ src/main/java/com/easypost/service/ScanformService.java
@@ -5,7 +5,7 @@ import com.easypost.http.Requestor;
 import com.easypost.http.Requestor.RequestMethod;
 import com.easypost.model.ScanForm;
 import com.easypost.model.ScanFormCollection;
-import com.easypost.utils.Utilities;
+import com.easypost.utils.InternalUtilities;
 
 import java.util.Map;
 
@@ -29,7 +29,7 @@ public class ScanformService {
      * @throws EasyPostException when the request fails. when the request fails.
      */
     public ScanForm create(final Map<String, Object> params) throws EasyPostException {
-        return Requestor.request(RequestMethod.POST, Utilities.classURL(ScanForm.class), params, ScanForm.class,
+        return Requestor.request(RequestMethod.POST, InternalUtilities.classURL(ScanForm.class), params, ScanForm.class,
                 client);
     }
 
@@ -41,8 +41,8 @@ public class ScanformService {
      * @throws EasyPostException when the request fails. when the request fails.
      */
     public ScanForm retrieve(final String id) throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.instanceURL(ScanForm.class, id), null, ScanForm.class,
-                client);
+        return Requestor.request(RequestMethod.GET, InternalUtilities.instanceURL(ScanForm.class, id), null,
+                ScanForm.class, client);
     }
 
     /**
@@ -54,7 +54,7 @@ public class ScanformService {
      */
     public ScanFormCollection all(final Map<String, Object> params)
             throws EasyPostException {
-        return Requestor.request(RequestMethod.GET, Utilities.classURL(ScanForm.class), params,
+        return Requestor.request(RequestMethod.GET, InternalUtilities.classURL(ScanForm.class), params,
                 ScanFormCollection.class, client);
     }
 }